pyproject.toml
[project]
name = "lipa-trust-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.0,<4.3.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
